Update MarkdownParse.java
Fixed the issues that were causing the failure of test-file7 and test-file8. Test-file7 was failing because the program did not break when a close bracket was not found, now it does. Test-file8 was failing because the code that checked for an exclamation mark was trying to access an invalid index, it no longer does that.